introduces graphql endpoint
