+++ server/world_server/schema.py
@@ -1,0 +1,10 @@
+import graphene
+
+import country.schema
+
+
+class Query(country.schema.Query, graphene.ObjectType):
+    pass
+
+
+schema = graphene.Schema(query=Query)
